Fix tests
* Remove duplicate babelrc file
* Remove ui test script that is no longer in use
* Consolidate jest setup files
* Remove babel-module-resolver
* Explicitly state packages using jest
* Add ".node" ext to jest file resolver
* Add display name to jest projects
* Auto reset mocks
