+++ jest.config.js
@@ -1,4 +1,4 @@
 module.exports = {
-  reporters: [ "default", "jest-junit" ],
-  projects: ["<rootDir>/packages/*"],
+  reporters: ["default", "jest-junit"],
+  projects: ["<rootDir>/packages/augur-node", "<rootDir>/packages/augur-ui"]
 };
