+++ package.json
@@ -20,7 +20,6 @@
     "docker:build:pops": "lerna run --scope augur-tools docker:build"
   },
   "devDependencies": {
-    "babel-plugin-module-resolver": "3.1.1",
     "babel-core": "7.0.0-bridge.0",
     "@babel/core": "7.1.6",
     "@babel/plugin-proposal-class-properties": "7.1.0",
