--- packages/augur-ui/test/setupFramework.js
@@ -1,6 +1,0 @@
-const localStorageMock = {
-  getItem: jest.fn(),
-  setItem: jest.fn(),
-  clear: jest.fn()
-};
-global.localStorage = localStorageMock;
