--- packages/augur.js/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    ["env",  {
-    "targets": {
-      "browsers": [
-        "last 1 version",
-        "> 3%"
-      ],
-      "node": "8"
-    }
-  }]]
-}
